Move GetDevices method
